requirements.txt
fastapi
uvicorn
jinja2
langchain
langchain-openai
langchain-groq
langgraph
python-dotenv
pandas
starlette
httpx